UART echoing works

+++ Code/Lab3_UART/uart.h
@@ -7,7 +7,7 @@
 
 #ifndef UART_H_
 #define UART_H_
-void config_uart(uint8_t baud);
+void config_uart(uint16_t baud);
 
 void enable_uart();
 void disable_uart();
